Moves InnerNavigation component to ui folder and remove it when is  not used
