+++ pages/advocates/index.vue
@@ -84,7 +84,7 @@
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
 import QiskitPage from '~/components/qiskit/QiskitPage.vue'
-import InnerNavigation from '~/components/menus/InnerNavigation.vue'
+import InnerNavigation from '~/components/ui/InnerNavigation.vue'
 import CommunityHeader from '~/components/headers/CommunityHeader.vue'
 import PageSection from '~/components/sections/PageSection.vue'
 import MapSection from '~/components/advocates/MapSection.vue'
