+++ pages/experiments/index.vue
@@ -36,7 +36,6 @@
 <script lang="ts">
 import { Component } from 'vue-property-decorator'
 import QiskitPage from '~/components/qiskit/QiskitPage.vue'
-import InnerNavigation from '~/components/menus/InnerNavigation.vue'
 import CommunityHeader from '~/components/headers/CommunityHeader.vue'
 import PageSection from '~/components/sections/PageSection.vue'
 import ExperimentCard from '~/components/experiments/ExperimentCard.vue'
@@ -44,7 +43,6 @@ import ExperimentCarousel from '~/components/experiments/ExperimentCarousel.vue'
 
 @Component({
   components: {
-    InnerNavigation,
     CommunityHeader,
     PageSection,
     ExperimentCard,
